+++ CHANGES.md
@@ -2,12 +2,14 @@ Changes
 =======
 0.4.0
 
-2013/10/09 pfraze
+2013/10/08 pfraze
 
+ - Added optional BridgeServer HOL blocking for unordered channels, toggleable with useMessageReordering()
+ - Updated RTCPeerServers to send HTTPL traffic over the Relay until the session is established
  - Changed the navigator `retry` param to `noretry`, making re-requests on failure the default
 
 
-2013/10/08 pfraze
+2013/10/07 pfraze
 
  - Renamed local.PeerWebRelay to local.Relay, local.joinPeerRelay to local.joinRelay
  - Added relay.navigator() and relay.registerLinks()